requirements.txt
pdf2image
gTTS
python-pptx
edge-tts
certifi
